requirements.txt
Django==1.1.1
django_extensions
django-taggit
sorl-thumbnail==3.2.5
django-piston
pil
django-haystack==1.1.0
django-memcached


-e git://github.com/flashingpumpkin/django-socialregistration.git#egg=socialregistration
-e git://github.com/facebook/python-sdk#egg=facebook
-e git://github.com/robhudson/django-debug-toolbar.git#egg=debug_toolbar
-e git://github.com/toastdriven/pysolr.git#egg=pysorl # 2.0.13